move to sched_rr - adapt to restricted cpuset
We want to shield a task from other tasks by using
cpuset. sched_deadline is a global scheduler that requires access to all
CPUs.

Moving to sched_rr and a fairly high prio. No kernel threads will run
with sched_deadline so this should reduce the interference and chance of
preemption to a minimum.